Group backup file types and parsing
